jackmatrix/requirements.txt
git+https://github.com/Rapptz/discord.py
async-timeout==3.0.1
python-dotenv==0.12.0
tabulate==0.8.6
flask==1.1.1